+++ src/syncData.ts
@@ -1,5 +1,6 @@
 import _ from 'lodash/fp.js'
 import type {
+  AnyBulkWriteOperation,
   ChangeStreamDocument,
   ChangeStreamInsertDocument,
   Collection,
@@ -28,23 +29,54 @@ export const initSync = (
     emitter.emit(event, { type: event, ...data })
   }
 
-  const processRecord = async (doc: ChangeStreamDocument) => {
+  /**
+   * Process change stream events.
+   */
+  const processChangeStreamRecords = async (docs: ChangeStreamDocument[]) => {
     try {
-      if (doc.operationType === 'insert') {
-        const document = mapper(doc.fullDocument)
-        await destination.insertOne(document)
-      } else if (
-        doc.operationType === 'update' ||
-        doc.operationType === 'replace'
-      ) {
-        const document = doc.fullDocument ? mapper(doc.fullDocument) : {}
-        await destination.replaceOne({ _id: doc.documentKey._id }, document, {
-          upsert: true,
-        })
-      } else if (doc.operationType === 'delete') {
-        await destination.deleteOne({ _id: doc.documentKey._id })
+      const operations: AnyBulkWriteOperation[] = []
+      for (const doc of docs) {
+        if (doc.operationType === 'insert') {
+          operations.push({
+            insertOne: {
+              document: mapper(doc.fullDocument),
+            },
+          })
+        } else if (
+          doc.operationType === 'update' ||
+          doc.operationType === 'replace'
+        ) {
+          const replacement = doc.fullDocument ? mapper(doc.fullDocument) : {}
+          operations.push({
+            replaceOne: {
+              filter: { _id: doc.documentKey._id },
+              replacement,
+              upsert: true,
+            },
+          })
+        } else if (doc.operationType === 'delete') {
+          operations.push({
+            deleteOne: {
+              filter: { _id: doc.documentKey._id },
+            },
+          })
+        }
       }
-      emit('process', { success: 1, changeStream: true })
+      const result = await destination.bulkWrite(operations, {
+        // Operations must be ordered
+        ordered: true,
+      })
+      const numSuccess = _.flow(
+        _.pick(['nInserted', 'nModified', 'nRemoved', 'nUpserted']),
+        Object.values,
+        _.sum
+      )(result)
+      const numFailed = operations.length - numSuccess
+      emit('process', {
+        success: numSuccess,
+        fail: numFailed,
+        changeStream: true,
+      })
     } catch (e) {
       emit('error', { error: e, changeStream: true })
     }
@@ -52,14 +84,15 @@ export const initSync = (
 
   const processRecords = async (docs: ChangeStreamInsertDocument[]) => {
     try {
-      const documents = docs.map(({ fullDocument }) => ({
+      const operations = docs.map(({ fullDocument }) => ({
         insertOne: { document: mapper(fullDocument) },
       }))
-      const result = await destination.bulkWrite(documents, { ordered: false })
-      const numInserted = result.nInserted
-      const numFailed = documents.length - numInserted
+      // Operations are unordered
+      const result = await destination.bulkWrite(operations, { ordered: false })
+      const numSuccess = result.nInserted
+      const numFailed = operations.length - numSuccess
       emit('process', {
-        success: numInserted,
+        success: numSuccess,
         fail: numFailed,
         initialScan: true,
       })
@@ -68,8 +101,8 @@ export const initSync = (
     }
   }
 
-  const processChangeStream = (options?: ChangeStreamOptions) =>
-    sync.processChangeStream(processRecord, options)
+  const processChangeStream = (options?: QueueOptions & ChangeStreamOptions) =>
+    sync.processChangeStream(processChangeStreamRecords, options)
   const runInitialScan = (options?: QueueOptions & ScanOptions) =>
     sync.runInitialScan(processRecords, options)
 
@@ -77,6 +110,8 @@ export const initSync = (
     ...sync,
     /**
      * Process MongoDB change stream for the given collection.
+     * `options.batchSize` defaults to 500.
+     * `options.timeout` defaults to 30 seconds.
      */
     processChangeStream,
     /**